TechShadow/requirements.txt
blinker==1.9.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
python-dotenv==1.0.1
tomli==2.1.0
Werkzeug==3.1.3
